innotter/requirements.txt
amqp==5.1.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
env==0.1.0
environs==9.5.0
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.0
marshmallow==3.15.0
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.28
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
psycopg2
boto3==1.21.23
django-filter==21.1
uuid==1.30
fastapi==0.75.0
pydantic==1.9.0
starlette==0.17.1
pika==1.2.0